+++ IKST_Workshop/Contents/mods/IKappaIDSuiteToolsTiles/42.18/media/lua/server/IKST_ProtectOps.lua
@@ -1,4 +1,8 @@
 -- Tile / vehicle protection server ops.
+if type(isClient) == "function" and isClient()
+    and type(isServer) == "function" and not isServer() then
+    return
+end
 
 require "IKST_Shared"
 require "IKST_TileProtect"